Extract workflows into a separate package
We now hackily set and absolute path to
/home/aiida/apps/aiidalab-ispg so that the package
can be build and installed automatically when AiiDAlab
is first launched. In the future we need to figure out
a better solution.
