+++ setup.cfg
@@ -33,10 +33,14 @@ python_requires = >=3.7
 # Once the version 1.7.2 is released, we should add it as a dependency to aiida-orca
 
 # xtb-python is published only via conda-forge so cannot be specified here :-(
+# The absolute path to aiidalab_atmospec_workchain is a hack for now,
+# at the end we will probably have to do the same thing as aiidalab-qe
+# and publich a separate package to PyPi.
 install_requires =
     aiidalab>=21.09.0
     cclib @ git+https://github.com/cclib/cclib.git@master
     aiida-orca @ git+https://github.com/pzarabadip/aiida-orca.git@master
+    aiidalab_atmospec_workchain @ file:///home/aiida/apps/aiidalab-ispg/workflows/
 
 [flake8]
 ignore =
