+++ workflows/setup.cfg
@@ -1,0 +1,33 @@
+[metadata]
+name = aiidalab_atmospec_workchain
+version = 0.1-alpha
+description = Package for the AiiDAlab Atmospec app workchain
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/danielhollas/aiidalab-ispg
+author = Daniel Hollas
+author_email = daniel.hollas@bristol.ac.uk
+classifiers =
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+project_urls =
+    Bug Tracker = https://github.com/danielhollas/aiidalab-ispg/issues
+
+[options]
+packages = find:
+# TODO: Add dependencies, although theu should be already satisfied
+# when this is being installed.
+#install_requires =
+python_requires = >=3.7
+
+[flake8]
+ignore =
+    E501  # Line length handled by black.
+    W503  # Line break before binary operator, preferred formatting for black.
+    E203  # Whitespace before ':', preferred formatting for black.
